requirements.txt
graphviz
networkx 
dot2tex 
pyyaml
click
